requirements.txt
dash
dash-renderer
dash-html-components
dash-core-components==0.24.0rc2
plotly
mysql-connector-python
sshtunnel